pyproject.toml
[tool.poetry]
name = "stream"
version = "0.1.0"
description = ""
authors = ["ngo quang dat <dat.ngoquang@glinteco.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
channels = {extras = ["daphne"], version = "^4.0.0"}
djangorestframework = "^3.14.0"
factory-boy = "^3.3.0"
faker = "^22.2.0"
coverage = "^7.4.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
)/
'''

[tool.isort]
profile = "black"
known_first_party = ["stream"]
src_paths = ["app"]
